Коммит 11 - create and refactor new Controllers
