+++ src/main/java/com/convenientservices/web/controllers/LoginController.java
@@ -1,0 +1,17 @@
+package com.convenientservices.web.controllers;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequiredArgsConstructor
+@RequestMapping("/login")
+public class LoginController {
+
+    @GetMapping()
+    public String showLoginPage() {
+        return "login";
+    }
+}
