+++ src/main/java/com/convenientservices/web/controllers/MainController.java
@@ -1,6 +1,5 @@
 package com.convenientservices.web.controllers;
 
-import com.convenientservices.web.entities.User;
 import com.convenientservices.web.services.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -35,31 +34,5 @@ public class MainController {
                                 Model model) {
         model.addAttribute("username", userService.getFIO(principal));
         return "about";
-    }
-
-    @GetMapping("/registration")
-    public String showRegistrationPage(Model model) {
-        model.addAttribute(new User());
-        model.addAttribute("answer", "");
-        return "registration";
-    }
-
-    @GetMapping("/login")
-    public String showLoginPage() {
-        return "login";
-    }
-
-    @PostMapping("/registration")
-    public String registrationUser(Model model,
-                                   @ModelAttribute("user") User user,
-                                   @RequestParam String role,
-                                   @RequestParam String matchingPassword) {
-        String answer = userService.registerNewUser(user, role, matchingPassword);
-        if ("success".equals(answer)) {
-            model.addAttribute("success", true);
-            return "login";
-        }
-        model.addAttribute("answer", answer);
-        return "registration";
     }
 }
